Refactor division and event services to include club validation and improve error handling; update repository queries for better data retrieval.

+++ backend/src/main/java/com/ovaltrack/backend/club/business/ClubService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import com.ovaltrack.backend.club.domain.Club;
 import com.ovaltrack.backend.club.repository.ClubRepository;
+import com.ovaltrack.backend.common.config.exceptions.BusinessException;
 
 import jakarta.transaction.Transactional;
 
@@ -37,8 +38,13 @@ public class ClubService {
     @Transactional
     public void deleteClub(UUID clubId) {
         //Verifications
+       if (clubRepository.findById(clubId) == null) {
+            throw new BusinessException("Club no encontrado");
+       }
         clubRepository.deleteById(clubId);
     }
 
 
 }
+
+
