+++ backend/src/main/java/com/ovaltrack/backend/match/business/MatchService.java
@@ -4,8 +4,11 @@ import java.util.Collection;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.ovaltrack.backend.common.config.exceptions.BusinessException;
 import com.ovaltrack.backend.division.domain.Division;
 import com.ovaltrack.backend.match.domain.Match;
 import com.ovaltrack.backend.match.repository.MatchRepository;
@@ -26,8 +29,8 @@ public class MatchService {
         return matchRepository.findAllMatchesByDivisionId(divisionId);
     }
 
-	public Match findMatchById(UUID matchId) {
-		return matchRepository.findById(matchId).orElse(null);
+	public Match findMatchByIdAndDivisionId(UUID matchId, UUID divisionId) {
+		return matchRepository.findMatchByIdAndDivisionId(matchId, divisionId);
 	}
 
 	@Transactional
@@ -37,7 +40,10 @@ public class MatchService {
 	}
 
 	@Transactional
-	public void deleteMatch(UUID matchId) {
+	public void deleteMatch(UUID divisionId, UUID matchId) {
+		if (findMatchByIdAndDivisionId(matchId, divisionId) == null) {
+			throw new BusinessException("No se puede eliminar un partido que no existe");
+		}
 		matchRepository.deleteById(matchId);
 	}
 }
